requirements.txt
Django==2.2
django-jet==1.0.8
django-multiselectfield==0.1.12
django-phone-field==1.8.1
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
